feat: Add MonitoraLicenca class
The code changes introduce a new file `monitoraLicenca.py` which contains the implementation of the `MonitoraLicenca` class. This class is currently empty, but it provides a foundation for monitoring licenses.

Note: The commit message follows the convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or functionality being added.