chore: Add SQL script to create database and tables
The code changes include the addition of a new SQL script `createDb.sql` which contains the necessary queries to create the database schema and tables for the `zanthus` schema. This script sets up the `atualizacao` and `monitoramento` tables with their respective columns and constraints.

Note: The commit message follows the convention of using a prefix to indicate the type of change. In this case, "chore" is used to indicate a routine task or maintenance change.