feat: Update config file path in AtualizaLicenca class
The code changes update the file path for the config file in the AtualizaLicenca class. The previous path "config/configHomolog.json" has been changed to "config/config.json".

Note: The commit message follows the convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or functionality being added.