feat: Update config file path and implement PyInstaller compilation
The code changes update the file path for the config file in the AtualizaLicenca class, changing it from "config/configHomolog.json" to "config/config.json". Additionally, the code changes include the implementation of PyInstaller compilation to create an executable file for the application.

Note: The commit message follows the convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or functionality being added.