Refactor database auto-increment values in createDb.sql
The code changes update the AUTO_INCREMENT values in the createDb.sql file for the `atualizacao` and `filiais` tables. The previous values have been changed to 1, which ensures that the first inserted record will have an ID of 1.

Note: The commit message follows the convention of using a prefix to indicate the type of change. In this case, "chore" could be used to indicate a miscellaneous or maintenance task.